serverless-deploy/test/requirements.txt
azure-storage-queue